[breaking] Implement co_awaiting QNetworkReply result in terms of QCoroNetworkReply
No changes in code are needed, only include headers must be adjusted to
include QCoroNetworkReply (qcoronetworkreply.h), QCoroNetwork
(qcoronetwork.h) or QCoro (qcoro.h) instead of network.h.

It is still possible to co_await QNetworkReply* directly (without the
qCoro wrapper), but that only internally delegates to a new
WaitForFinishedOperation.
